Refactoring folder structure and submit button style modification
